Fix parsing of Xresources
There doesn't need to be a dot before the identifier, so we can have any
of:

    *.color0: #333333
    *color0: #333333
    Some.Path.color0: #333333

Previously the regex was depending on a dot being present. We could
extract the label using a single regex if we did a bunch of backtracking
or used a non-greedy quantifier, but I think it's clearer to a
substitution on the captured label instead.